backend/requirements.txt
# FastAPI and core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0

# For image processing
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0

# Pydantic for data validation
pydantic>=2.0.0

# CORS
starlette>=0.27.0

# AI/ML dependencies (already in parent requirements)
torch
diffusers
transformers
accelerate
scipy
